+++ L13-structuralPatterns/src/main/java/ru/otus/ATMain.java
@@ -6,11 +6,11 @@ import ru.otus.banknotes.*;
 import ru.otus.banknotes.adapter.*;
 
 public class ATMain {
-    public static void main(String[] args){
+    public static void main(String[] args) {
         checkObserverWithMemento();
     }
 
-    public static void checkAdapter(){
+    public static void checkAdapter() {
         BundleOfBanknotes bundle = new BundleOfBanknotes();
         System.out.println(bundle);
         bundle.addBanknotes(new BanknoteImpl(Currency.USD, Nominal.FIFTY), 11);
@@ -21,29 +21,29 @@ public class ATMain {
         System.out.println(bundle);
     }
 
-    public static void checkCommand(){
-        AbstractATM firstAtm = new ATMImpl();
-        AbstractATM secondAtm = new ATMImpl();
+    public static void checkCommand() {
+        ATM firstAtm = new ATMImpl();
+        ATM secondAtm = new ATMImpl();
 
         DepartmentOfATM department = new DepartmentOfATM();
         department.addATM(firstAtm);
         department.addATM(secondAtm);
 
-        firstAtm.loadCassette(new Cassette(new BanknoteImpl(Currency.USD,Nominal.HUNDRED)));
-        secondAtm.loadCassette(new Cassette(new BanknoteImpl(Currency.USD,Nominal.HUNDRED)));
+        firstAtm.loadCassette(new CassetteImpl(new BanknoteImpl(Currency.USD, Nominal.HUNDRED)));
+        secondAtm.loadCassette(new CassetteImpl(new BanknoteImpl(Currency.USD, Nominal.HUNDRED)));
 
         department.onAllAtm();
-        secondAtm.loadCassette(new Cassette(new BanknoteImpl(Currency.USD,Nominal.HUNDRED)));
+        secondAtm.loadCassette(new CassetteImpl(new BanknoteImpl(Currency.USD, Nominal.HUNDRED)));
 
         department.offAllAtm();
-        secondAtm.loadCassette(new Cassette(new BanknoteImpl(Currency.USD,Nominal.HUNDRED)));
+        secondAtm.loadCassette(new CassetteImpl(new BanknoteImpl(Currency.USD, Nominal.HUNDRED)));
     }
 
-    public static void checkObserverWithMemento(){
+    public static void checkObserverWithMemento() {
         DepartmentOfATM department = new DepartmentOfATM();
-        AbstractATM atm1 = getDefaultATM();
-        AbstractATM atm2 = getDefaultATM();
-        AbstractATM atm3 = getDefaultATM();
+        ATM atm1 = getDefaultATM();
+        ATM atm2 = getDefaultATM();
+        ATM atm3 = getDefaultATM();
         department.addATM(atm1);
         department.addATM(atm2);
         department.addATM(atm3);
@@ -53,30 +53,23 @@ public class ATMain {
         System.out.println("Баланс рублей в atm1 после взноса = " + atm1.getBalance(Currency.RUB));
         atm1.cashOut(new BanknoteImpl(Currency.RUB, Nominal.FIFTY), 600);
         System.out.println("Баланс рублей в atm1 после выдачи = " + atm1.getBalance(Currency.RUB));
-        //atm2.cashOut(new BanknoteImpl(Currency.RUB, Nominal.FIFTY), 100);
-
-        //department.returnToStartState();
-        //System.out.println("Баланс рублей в atm1 после сброса = " + atm1.getBalance(Currency.RUB));
-
         System.out.println(department.getBalanceSummary());
-
     }
 
+    private static ATM getDefaultATM() {
 
-    private static ATMImpl getDefaultATM(){
-
-        Cassette cassetteRub = new Cassette(new BanknoteImpl(Currency.RUB, Nominal.FIFTY));
+        CassetteImpl cassetteRub = new CassetteImpl(new BanknoteImpl(Currency.RUB, Nominal.FIFTY));
         cassetteRub.addBanknotes(30);
 
-        Cassette cassetteRubSecond = new Cassette(new BanknoteImpl(Currency.RUB, Nominal.HUNDRED));
+        CassetteImpl cassetteRubSecond = new CassetteImpl(new BanknoteImpl(Currency.RUB, Nominal.HUNDRED));
 
-        Cassette cassetteEmpire = new Cassette(new BanknoteImpl(Currency.EUR, Nominal.HUNDRED));
+        CassetteImpl cassetteEmpire = new CassetteImpl(new BanknoteImpl(Currency.EUR, Nominal.HUNDRED));
         cassetteEmpire.addBanknotes(6);
 
-        Cassette cassetteUsd = new Cassette(new BanknoteImpl(Currency.USD, Nominal.FIFTY));
+        CassetteImpl cassetteUsd = new CassetteImpl(new BanknoteImpl(Currency.USD, Nominal.FIFTY));
         cassetteUsd.addBanknotes(16);
 
-        ATMImpl atm = new ATMImpl();
+        ATM atm = new ATMImpl();
         atm.on();
         atm.setCashOutBehavior(new BanknoteBehavior());
         atm.loadCassette(cassetteRub);
