+++ L13-structuralPatterns/src/main/java/ru/otus/atm/ATMImpl.java
@@ -1,14 +1,68 @@
 package ru.otus.atm;
 
+import ru.otus.atm.memento.Memento;
+import ru.otus.atm.observer.Directives;
+import ru.otus.atm.observer.Listener;
 import ru.otus.atm.strategy.BanknoteBehavior;
 import ru.otus.atm.decorator.ComissionForNotRub;
 import ru.otus.atm.decorator.ComissionForOverSumm;
 import ru.otus.atm.decorator.DefaultChanger;
+import ru.otus.atm.strategy.CashOutBehavior;
 import ru.otus.banknotes.*;
 
-public class ATMImpl extends AbstractATM {
+import java.io.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+public class ATMImpl implements ATM, Listener, Serializable {
+    private List<CassetteImpl> cassettes = new ArrayList<>();
+    private CashOutBehavior cashOutBehavior;
+    private int workState = 0;
+    private Memento undo = null;
+
     public ATMImpl(){
         cashOutBehavior = new BanknoteBehavior();
+    }
+
+    public void loadCassette(CassetteImpl cassette){
+        cassettes.add(cassette);
+        System.out.println("Кассета установлена");
+    }
+
+    public void off() {
+        workState = 0;
+    }
+
+    public void on() {
+        workState = 1;
+    }
+
+    public Memento getUndo() {
+        return undo;
+    }
+
+    public int getWorkState() {
+        return workState;
+    }
+
+    public List<CassetteImpl> getCassetes() {
+        return cassettes;
+    }
+
+    public CashOutBehavior getCashOutBehavior() {
+        return cashOutBehavior;
+    }
+
+    public void setCashOutBehavior(CashOutBehavior cashOutBehavior){
+        this.cashOutBehavior = cashOutBehavior;
+    }
+
+    public int getBalance(Currency currency){
+        return cassettes.stream()
+                .filter(cassette -> cassette.getUsedBanknote().getCurrency() == currency)
+                .mapToInt(CassetteImpl::getBalance).sum();
     }
 
     @Override
@@ -38,4 +92,105 @@ public class ATMImpl extends AbstractATM {
             }
         }
     }
+
+    public void cashIn(Banknote banknote, int count){
+        if(!isWorkAtm())return;
+        if(
+                cassettes.stream()
+                        .filter(cassette -> cassette.getUsedBanknote().equals(banknote))
+                        .mapToInt(x->x.getMaxCountBanknotes() - x.getCountBanknotes())
+                        .sum() < count
+        ) {
+            System.out.println("Нужные кассеты не установлены или в них не хватит места");
+        } else {
+            saveState();
+            int countForAdd = count;
+            for(CassetteImpl cassette: cassettes){
+                if(cassette.getUsedBanknote().equals(banknote)){
+                    countForAdd -= cassette.addBanknotesToMax(countForAdd);
+                    if(countForAdd <= 0) break;
+                }
+            }
+        }
+    }
+
+    public void returnToStartState(){
+        while (undo != null){
+            try {
+                ByteArrayInputStream bis = new ByteArrayInputStream(undo.getState());
+                ObjectInputStream ois = new ObjectInputStream(bis);
+                ATM restoreAtm = (ATM) ois.readObject();
+                this.undo = restoreAtm.getUndo();
+                this.workState = restoreAtm.getWorkState();
+                this.cassettes = restoreAtm.getCassetes();
+                this.cashOutBehavior = restoreAtm.getCashOutBehavior();
+            } catch (IOException | ClassNotFoundException e) {
+                e.printStackTrace();
+                break;
+            }
+        }
+    }
+
+    public Map<Currency, Integer> getBalance(){
+        Map<Currency, Integer> result = new HashMap<>();
+        cassettes.forEach(cassette -> {
+            if(cassette.getCountBanknotes() > 0){
+                result.put(
+                        cassette.getUsedBanknote().getCurrency(),
+                        result.getOrDefault(cassette.getUsedBanknote().getCurrency(), 0)
+                                + cassette.getCountBanknotes() * cassette.getUsedBanknote().getNominal().getValue());
+            }
+        });
+        return result;
+    }
+
+    @Override
+    public void executeDirective(Directives directive) {
+        if(directive == Directives.ON){
+            on();
+        } else if(directive == Directives.OFF){
+            off();
+        } else if(directive == Directives.RESET){
+            returnToStartState();
+        }
+    }
+
+    private boolean isWorkAtm(){
+        if(workState == 1){
+            return true;
+        } else {
+            System.out.println("Банкомат не работает");
+            return false;
+        }
+    }
+
+    private void saveState(){
+        try {
+            ByteArrayOutputStream bos = new ByteArrayOutputStream();
+            ObjectOutputStream oos = new ObjectOutputStream(bos);
+            oos.writeObject(this);
+            undo = new Memento(bos.toByteArray());
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    private int getCountBanknotes(Banknote banknote){
+        return cassettes.stream()
+                .filter(cassette -> cassette.getUsedBanknote().equals(banknote))
+                .mapToInt(CassetteImpl::getCountBanknotes)
+                .sum();
+    }
+
+    private void recalculateCassettes(BundleOfBanknotes cashOutBundle){
+        cashOutBundle.getBanknotes().forEach((banknote, count) -> {
+            for(CassetteImpl cassette: cassettes){
+                if(cassette.getUsedBanknote().equals(banknote) && cassette.getCountBanknotes() > 0){
+                    count -= cassette.removeBanknotesToMin(count);
+                    if(count <= 0) break;
+                }
+            }
+        });
+    }
+
 }
