+++ L13-structuralPatterns/src/main/java/ru/otus/atm/Cassette.java
@@ -1,66 +1,21 @@
 package ru.otus.atm;
 
-import ru.otus.banknotes.*;
+import ru.otus.banknotes.Banknote;
 
-public class Cassette {
-    private Banknote banknote;
-    private BundleOfBanknotes bundleOfBanknotes = new BundleOfBanknotes();
-    private int maxCountBanknotes = 50;
+public interface Cassette {
+    public int getMaxCountBanknotes();
 
-    public Cassette(BanknoteImpl banknoteImpl){
-        this.banknote = banknoteImpl;
-    }
+    public int getCountBanknotes();
 
-    public Cassette(BanknoteImpl banknote, int maxCount){
-        this.banknote = banknote;
-        maxCountBanknotes = maxCount;
-    }
+    public void addBanknotes(int countAddBanknotes);
 
-    public int getMaxCountBanknotes(){
-        return maxCountBanknotes;
-    }
+    public int addBanknotesToMax(int countAddBanknotes);
 
-    public int getCountBanknotes() {
-        return bundleOfBanknotes.getCountBanknotes(banknote);
-    }
+    public void removeBanknotes(int countRemoveBanknotes);
 
-    public void addBanknotes(int countAddBanknotes){
-        if(bundleOfBanknotes.getCountBanknotes() + countAddBanknotes > maxCountBanknotes) throw new RuntimeException("Кассета переполнится");
-        this.bundleOfBanknotes.addBanknotes(banknote, countAddBanknotes);
-    }
+    public int removeBanknotesToMin(int countRemoveBanknotes);
 
-    public int addBanknotesToMax(int countAddBanknotes){
-        int countFree = maxCountBanknotes - bundleOfBanknotes.getCountBanknotes();
-        if(countFree >= countAddBanknotes){
-            addBanknotes(countAddBanknotes);
-            return countAddBanknotes;
-        } else {
-            addBanknotes(countFree);
-            return countFree;
-        }
-    }
+    public int getBalance();
 
-    public void removeBanknotes(int countRemoveBanknotes){
-        if(bundleOfBanknotes.getCountBanknotes() - countRemoveBanknotes < 0) throw new RuntimeException("В кассете не хватает банкнот");
-        bundleOfBanknotes.removeBanknotes(banknote, countRemoveBanknotes);
-    }
-
-    public int removeBanknotesToMin(int countRemoveBanknotes){
-        if(getCountBanknotes() >= countRemoveBanknotes){
-            removeBanknotes(countRemoveBanknotes);
-            return countRemoveBanknotes;
-        } else {
-            int countRemoved = getCountBanknotes();
-            removeBanknotes(countRemoved);
-            return countRemoved;
-        }
-    }
-
-    public int getBalance(){
-        return bundleOfBanknotes.getCountBanknotes(banknote) * banknote.getNominal().getValue();
-    }
-
-    public Banknote getUsedBanknote(){
-        return banknote;
-    }
+    public Banknote getUsedBanknote();
 }
