+++ L13-structuralPatterns/src/main/java/ru/otus/atm/commands/OffAllAtmCommand.java
@@ -1,19 +1,19 @@
 package ru.otus.atm.commands;
 
-import ru.otus.atm.AbstractATM;
+import ru.otus.atm.ATM;
 
 import java.util.Collection;
 
 public class OffAllAtmCommand implements Command {
-    private final Collection<AbstractATM> atms;
+    private final Collection<ATM> atms;
 
-    public OffAllAtmCommand(Collection<AbstractATM> atms){
+    public OffAllAtmCommand(Collection<ATM> atms){
         this.atms = atms;
     }
 
     @Override
     public void execute() {
-        for(AbstractATM atm : atms){
+        for(ATM atm : atms){
             atm.off();
         }
         System.out.println("Все банкоматы выключены");
