+++ L13-structuralPatterns/src/main/java/ru/otus/atm/strategy/BanknoteBehavior.java
@@ -1,24 +1,25 @@
 package ru.otus.atm.strategy;
 
-import ru.otus.atm.Cassette;
+import ru.otus.atm.CassetteImpl;
 import ru.otus.banknotes.*;
 
+import java.io.Serializable;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class BanknoteBehavior implements CashOutBehavior {
+public class BanknoteBehavior implements CashOutBehavior, Serializable {
     @Override
-    public BundleOfBanknotes getBundleToCashOut(Currency currency, int summ, List<Cassette> cassettes) {
+    public BundleOfBanknotes getBundleToCashOut(Currency currency, int summ, List<CassetteImpl> cassettes) {
         cassettes.sort(Comparator.comparing(cassette -> cassette.getUsedBanknote().getNominal().getValue()));
-        List<Cassette> mutableCassettes = cassettes.stream()
+        List<CassetteImpl> mutableCassettes = cassettes.stream()
                 .filter(cassette -> cassette.getUsedBanknote().getCurrency() == currency)
                 .sorted(Comparator.comparing(cassette -> cassette.getUsedBanknote().getNominal().getValue()))
                 .collect(Collectors.toList());
         Collections.reverse(mutableCassettes);
         BundleOfBanknotes resultBundle = new BundleOfBanknotes();
-        for(Cassette cassette: mutableCassettes){
+        for(CassetteImpl cassette: mutableCassettes){
             int countInCassette = cassette.getCountBanknotes();
             while (countInCassette > 0 && cassette.getUsedBanknote().getNominal().getValue() <= summ && summ > 0){
                 resultBundle.addBanknotes(cassette.getUsedBanknote(), 1);
