+++ L13-structuralPatterns/src/main/java/ru/otus/atm/strategy/CashOutBehavior.java
@@ -1,10 +1,10 @@
 package ru.otus.atm.strategy;
 
-import ru.otus.atm.Cassette;
+import ru.otus.atm.CassetteImpl;
 import ru.otus.banknotes.*;
 
 import java.util.List;
 
 public interface CashOutBehavior {
-    public BundleOfBanknotes getBundleToCashOut(Currency currency, int summ, List<Cassette> cassettes);
+    public BundleOfBanknotes getBundleToCashOut(Currency currency, int summ, List<CassetteImpl> cassettes);
 }
